refactor: 부스 좋아요 기능 관련 훅 및 API 모듈 구조 리팩토링
