+++ src/api/booth/user/hooks/useUserBoothPubStatus.ts
@@ -7,7 +7,7 @@
 
 import { useQuery } from "@tanstack/react-query";
 import { fetchPubsStatus } from "../userBooth";
-import type { PubStatus } from "../../shared/sharedTypes";
+import type { PubStatus } from "../../shared/sharedBoothTypes";
 
 export const usePubStatus = (boothId: string) => {
   return useQuery<PubStatus | null>({
