+++ src/api/likes/likeBoothTypes.ts
@@ -1,0 +1,13 @@
+export interface LikeBoothResponse {
+  booth_id: number;
+  new_likes: number;
+}
+
+export interface LikeTotalResponse {
+  success: boolean;
+  data: {
+    booth_id: number;
+    new_likes: number;
+  };
+  error: unknown;
+}
