+++ src/pages/waiting/WaitingPage.tsx
@@ -10,8 +10,8 @@ import QuestionIcon from "@/assets/icons/question.png";
 import TopBar from "@/components/topbar/TopBar";
 // import { requestPermissionAndGetToken } from "@/firebase";
 import { cancelReservation, fetchMyReservation } from "@/api/booth/user/userBooth";
-import { useMyReservation } from "@/api/booth/user/hooks/useMyReservation";
-import { usePubStatuses } from "@/api/booth/user/hooks/usePubStatuses";
+import { useMyReservation } from "@/api/booth/user/hooks/useUserBoothReservation";
+import { usePubStatuses } from "@/api/booth/user/hooks/useUserBoothPubStatuses";
 
 const today = (() => {
   const now = new Date();
